Upgrded to fmt version 5.2.1
